+++ src/platform_impl/linux/x11/event_processor.rs
@@ -956,10 +956,22 @@ impl<T: 'static> EventProcessor<T> {
                         let dpi_factor =
                             self.with_window(xev.event, |window| window.hidpi_factor());
                         if let Some(dpi_factor) = dpi_factor {
+                            let modifiers = self.device_mod_state.modifiers();
                             let location = LogicalPosition::from_physical(
                                 (xev.event_x as f64, xev.event_y as f64),
                                 dpi_factor,
                             );
+
+                            // Mouse cursor position changes when touch events are received
+                            callback(Event::WindowEvent {
+                                window_id,
+                                event: WindowEvent::CursorMoved {
+                                    device_id: mkdid(util::VIRTUAL_CORE_POINTER),
+                                    position: location,
+                                    modifiers,
+                                },
+                            });
+
                             callback(Event::WindowEvent {
                                 window_id,
                                 event: WindowEvent::Touch(Touch {