+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - On X11, fix key modifiers being incorrectly reported.
 - On X11, fix window creation hanging when another window is fullscreen.
 - On Windows, fix focusing unfocused windows when switching from fullscreen to windowed.
+- On X11, issue a `CursorMoved` event when a `Touch` event occurs,
+  as X11 implicitly moves the cursor for such events.
 
 # 0.20.0 Alpha 4 (2019-10-18)